+++ src/dipdup/datasources/tzkt/datasource.py
@@ -40,6 +40,7 @@ from dipdup.datasources.subscription import BigMapSubscription
 from dipdup.datasources.subscription import HeadSubscription
 from dipdup.datasources.subscription import OriginationSubscription
 from dipdup.datasources.subscription import Subscription
+from dipdup.datasources.subscription import TokenTransferSubscription
 from dipdup.datasources.subscription import TransactionSubscription
 from dipdup.datasources.tzkt.enums import ORIGINATION_MIGRATION_FIELDS
 from dipdup.datasources.tzkt.enums import ORIGINATION_OPERATION_FIELDS
@@ -47,6 +48,7 @@ from dipdup.datasources.tzkt.enums import TRANSACTION_OPERATION_FIELDS
 from dipdup.datasources.tzkt.enums import OperationFetcherRequest
 from dipdup.datasources.tzkt.enums import TzktMessageType
 from dipdup.enums import MessageType
+from dipdup.enums import TokenStandard
 from dipdup.exceptions import DatasourceError
 from dipdup.models import BigMapAction
 from dipdup.models import BigMapData
@@ -54,6 +56,7 @@ from dipdup.models import BlockData
 from dipdup.models import HeadBlockData
 from dipdup.models import OperationData
 from dipdup.models import QuoteData
+from dipdup.models import TokenTransferData
 from dipdup.utils import FormattedLogger
 from dipdup.utils import split_by_chunks
 from dipdup.utils.watchdog import Watchdog
@@ -260,6 +263,48 @@ class BigMapFetcher:
 
         if big_maps:
             yield big_maps[0].level, big_maps
+
+
+class TokenTransferFetcher:
+    def __init__(
+        self,
+        datasource: 'TzktDatasource',
+        first_level: int,
+        last_level: int,
+        cache: bool = False,
+    ) -> None:
+        self._logger = logging.getLogger('dipdup.tzkt')
+        self._datasource = datasource
+        self._first_level = first_level
+        self._last_level = last_level
+        self._cache = cache
+
+    async def fetch_token_transfers_by_level(self) -> AsyncGenerator[Tuple[int, Tuple[TokenTransferData, ...]], None]:
+        token_transfers: Tuple[TokenTransferData, ...] = ()
+
+        # TODO: Share code between this and OperationFetcher
+        token_transfer_iter = self._datasource.iter_token_transfers(
+            self._first_level,
+            self._last_level,
+        )
+        async for fetched_token_transfers in token_transfer_iter:
+            token_transfers = token_transfers + fetched_token_transfers
+
+            # NOTE: Yield token transfer slices by level except the last one
+            while True:
+                for i in range(len(token_transfers) - 1):
+                    curr_level, next_level = token_transfers[i].level, token_transfers[i + 1].level
+
+                    # NOTE: Level boundaries found. Exit for loop, stay in while.
+                    if curr_level != next_level:
+                        yield curr_level, token_transfers[: i + 1]
+                        token_transfers = token_transfers[i + 1 :]
+                        break
+                else:
+                    break
+
+        if token_transfers:
+            yield token_transfers[0].level, token_transfers
 
 
 MessageData = Union[Dict[str, Any], List[Dict[str, Any]]]
@@ -726,6 +771,38 @@ class TzktDatasource(IndexDatasource):
         ):
             yield batch
 
+    async def get_token_transfers(
+        self,
+        first_level: int,
+        last_level: int,
+        offset: Optional[int] = None,
+        limit: Optional[int] = None,
+    ) -> Tuple[TokenTransferData, ...]:
+        """Get token transfers for contract"""
+        offset, limit = offset or 0, limit or self.request_limit
+        params: dict = {}
+
+        raw_token_transfers = await self.request(
+            'get',
+            url='v1/tokens/transfers',
+            params={**params, 'level.ge': first_level, 'level.lt': last_level, 'offset': offset, 'limit': limit, 'sort.asc': 'level'},
+        )
+        return tuple(self.convert_token_transfer(item) for item in raw_token_transfers)
+
+    async def iter_token_transfers(
+        self,
+        first_level: int,
+        last_level: int,
+    ) -> AsyncIterator[Tuple[TokenTransferData, ...]]:
+        """Iterate token transfers for contract"""
+        async for batch in self._iter_batches(
+            self.get_token_transfers,
+            first_level,
+            last_level,
+            cursor=False,
+        ):
+            yield batch
+
     async def add_index(self, index_config: ResolvedIndexConfigT) -> None:
         """Register index config in internal mappings and matchers. Find and register subscriptions."""
         for subscription in index_config.subscriptions:
@@ -766,6 +843,10 @@ class TzktDatasource(IndexDatasource):
                 request = [{}]
             else:
                 raise RuntimeError
+
+        elif isinstance(subscription, TokenTransferSubscription):
+            method = 'SubscribeToTokenTransfers'
+            request = [{}]
 
         else:
             raise NotImplementedError
@@ -1087,6 +1168,35 @@ class TzktDatasource(IndexDatasource):
             jpy=Decimal(quote_json['jpy']),
             krw=Decimal(quote_json['krw']),
             eth=Decimal(quote_json['eth']),
+        )
+
+    @classmethod
+    def convert_token_transfer(cls, token_transfer_json: Dict[str, Any]) -> TokenTransferData:
+        """Convert raw token transfer message from REST or WS into dataclass"""
+        token_json = token_transfer_json.get('token') or {}
+        contract_json = token_json.get('contract') or {}
+        from_json = token_transfer_json.get('from') or {}
+        to_json = token_transfer_json.get('to') or {}
+        standard = token_json.get('standard')
+        metadata = token_json.get('metadata')
+        return TokenTransferData(
+            id=token_transfer_json['id'],
+            level=token_transfer_json['level'],
+            timestamp=cls._parse_timestamp(token_transfer_json['timestamp']),
+            tzkt_token_id=token_json['id'],
+            contract_address=contract_json.get('address'),
+            contract_alias=contract_json.get('alias'),
+            token_id=token_json.get('tokenId'),
+            standard=TokenStandard(standard) if standard else None,
+            metadata=metadata if isinstance(metadata, dict) else {},
+            from_alias=from_json.get('alias'),
+            from_address=from_json.get('address'),
+            to_alias=to_json.get('alias'),
+            to_address=to_json.get('address'),
+            amount=token_transfer_json.get('amount'),
+            tzkt_transaction_id=token_transfer_json.get('transactionId'),
+            tzkt_origination_id=token_transfer_json.get('originationId'),
+            tzkt_migration_id=token_transfer_json.get('migrationId'),
         )
 
     async def _send(