+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## [unreleased]
 
+### Added
+
+* ci: Push `X` and `X.Y` tags to the Docker Hub on release.
+* cli: Added `config env` command to export env-file with default values.
+* cli: Show warning when running an outdated version of DipDup.
+
 ### Fixed
 
 * index: Fixed fetching `migration` operations.